커뮤니티 -여행후기 main화면 '좋아요'순 정렬 추가

+++ tourGo/src/main/java/org/tourGo/models/community/review/LikedEntityRepository.java
@@ -7,5 +7,5 @@ import org.tourGo.models.entity.community.review.ReviewEntity;
 
 public interface LikedEntityRepository extends JpaRepository<LikedEntity, Long>, QuerydslPredicateExecutor {
 
-	long countByReviewAndLiked(ReviewEntity review, boolean liked);
+	int countByReviewAndLiked(ReviewEntity review, boolean liked);
 }
